+++ test/reach/project/project_test.exs
@@ -1,7 +1,9 @@
 defmodule Reach.ProjectTest do
   use ExUnit.Case, async: false
 
+  alias Reach.CLI.Project, as: CLIProject
   alias Reach.Project
+  alias Reach.Project.Query
 
   @tmp_dir Path.join(
              System.tmp_dir!(),
@@ -108,12 +110,12 @@ defmodule Reach.ProjectTest do
       path_a = write_file("lib/cache_a.ex", "defmodule CacheA do\n  def only_a, do: 1\nend\n")
       path_b = write_file("lib/cache_b.ex", "defmodule CacheB do\n  def only_b, do: 2\nend\n")
 
-      project_a = Reach.CLI.Project.load(paths: [path_a], quiet: true)
-      assert Reach.CLI.Project.resolve_target(project_a, "only_a/0") == {nil, :only_a, 0}
+      project_a = CLIProject.load(paths: [path_a], quiet: true)
+      assert Query.resolve_target(project_a, "only_a/0") == {nil, :only_a, 0}
 
-      project_b = Reach.CLI.Project.load(paths: [path_b], quiet: true)
-      assert Reach.CLI.Project.resolve_target(project_b, "only_b/0") == {nil, :only_b, 0}
-      assert Reach.CLI.Project.resolve_target(project_b, "only_a/0") == nil
+      project_b = CLIProject.load(paths: [path_b], quiet: true)
+      assert Query.resolve_target(project_b, "only_b/0") == {nil, :only_b, 0}
+      assert Query.resolve_target(project_b, "only_a/0") == nil
     end
   end